Conver robust to TypeScript
